+++ src/feditest/nodedrivers/imp/__init__.py
@@ -2,8 +2,6 @@
 An in-process Node implementation for now.
 """
 
-from typing import cast
-
 import httpx
 from multidict import MultiDict
 
@@ -14,19 +12,17 @@ from feditest.protocols.web.diag import (
     HttpResponse,
     WebDiagClient
 )
-from feditest.protocols.webfinger import WebFingerServer
-from feditest.protocols.webfinger.diag import ClaimedJrd, WebFingerDiagClient, WebFingerQueryResponse
-from feditest.protocols.webfinger.utils import construct_webfinger_uri_for
+from feditest.protocols.webfinger.abstract import AbstractWebFingerDiagClient
 from feditest.reporting import trace
 from feditest.testplan import TestPlanConstellationNode, TestPlanNodeParameter
-from feditest.utils import FEDITEST_VERSION, ParsedUri
+from feditest.utils import FEDITEST_VERSION
 
 _HEADERS = {
     "User-Agent": f"feditest/{ FEDITEST_VERSION }",
     "Origin": "test.example" # to trigger CORS headers in response
 }
 
-class Imp(WebFingerDiagClient):
+class Imp(AbstractWebFingerDiagClient):
     """
     In-process diagnostic WebFinger client.
     """
@@ -50,70 +46,6 @@ class Imp(WebFingerDiagClient):
             trace( f'HTTP query returns { ret }')
             return ret
         raise WebDiagClient.HttpUnsuccessfulError(request)
-
-
-    # Python 3.12 @override
-    def diag_perform_webfinger_query(
-        self,
-        resource_uri: str,
-        rels: list[str] | None = None,
-        server: WebFingerServer | None = None
-    ) -> WebFingerQueryResponse:
-        query_url = construct_webfinger_uri_for(resource_uri, rels, server.hostname() if server else None )
-        parsed_uri = ParsedUri.parse(query_url)
-        if not parsed_uri:
-            raise ValueError('Not a valid URI:', query_url) # can't avoid this
-        first_request = HttpRequest(parsed_uri)
-        current_request = first_request
-        pair : HttpRequestResponsePair | None = None
-        for redirect_count in range(10, 0, -1):
-            pair = self.http(current_request)
-            if pair.response and pair.response.is_redirect():
-                if redirect_count <= 0:
-                    return WebFingerQueryResponse(pair, None, WebDiagClient.TooManyRedirectsError(current_request))
-                parsed_location_uri = ParsedUri.parse(pair.response.location())
-                if not parsed_location_uri:
-                    return WebFingerQueryResponse(pair, None, ValueError('Location header is not a valid URI:', query_url, '(from', resource_uri, ')'))
-                current_request = HttpRequest(parsed_location_uri)
-            break
-
-        # I guess we always have a non-null responses here, but mypy complains without the cast
-        pair = cast(HttpRequestResponsePair, pair)
-        ret_pair = HttpRequestResponsePair(first_request, current_request, pair.response)
-        if ret_pair.response is None:
-            raise RuntimeError('Unexpected None HTTP response')
-
-        excs : list[Exception] = []
-        if ret_pair.response.http_status != 200:
-            excs.append(WebDiagClient.WrongHttpStatusError(ret_pair))
-
-        content_type = ret_pair.response.content_type()
-        if (content_type is None or (content_type != "application/jrd+json"
-            and not content_type.startswith( "application/jrd+json;" ))
-        ):
-            excs.append(WebDiagClient.WrongContentTypeError(ret_pair))
-
-        jrd : ClaimedJrd | None = None
-
-        if ret_pair.response.payload is None:
-            raise RuntimeError('Unexpected None payload in HTTP response')
-
-        try:
-            json_string = ret_pair.response.payload.decode(encoding=ret_pair.response.payload_charset() or "utf8")
-
-            jrd = ClaimedJrd(json_string) # May throw JSONDecodeError
-            jrd.validate() # May throw JrdError
-        except ExceptionGroup as exc:
-            excs += exc.exceptions
-        except Exception as exc:
-            excs.append(exc)
-
-        if len(excs) > 1:
-            return WebFingerQueryResponse(ret_pair, jrd, ExceptionGroup('WebFinger errors', excs))
-        elif len(excs) == 1:
-            return WebFingerQueryResponse(ret_pair, jrd, excs[0])
-        else:
-            return WebFingerQueryResponse(ret_pair, jrd, None)
 
 
     # Python 3.12 @override
